Fix compiler warning in dprint() (closes #3120)

+++ pager.c
@@ -1104,7 +1104,7 @@ static int format_line (struct line_t **lineInfo, int n, unsigned char *buf,
 
     if (Charset_is_utf8 && (wc == 0x200B || wc == 0xFEFF))
     {
-	dprint (3, (debugfile, "skip zero-width character U+%04X\n", wc));
+	dprint (3, (debugfile, "skip zero-width character U+%04X\n", (unsigned short)wc));
 	continue;
     }
